+++ esp_video/examples/common_components/example_video_common/example_init_video.c
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: ESPRESSIF MIT
  */
 
+#include <string.h>
 #include "esp_log.h"
 #include "esp_check.h"
 #include "example_video_common.h"
@@ -77,36 +78,72 @@ static const esp_video_init_dvp_config_t s_dvp_config = {
 #endif /* EXAMPLE_ENABLE_DVP_CAM_SENSOR */
 
 #if EXAMPLE_ENABLE_SPI_CAM_SENSOR
-static const esp_video_init_spi_config_t s_spi_config = {
-    .sccb_config = {
+static const esp_video_init_spi_config_t s_spi_config[] = {
+    {
+        .sccb_config = {
 #if !CONFIG_EXAMPLE_SCCB_I2C_INIT_BY_APP
-        .init_sccb = true,
-        .i2c_config = {
-            .port      = EXAMPLE_SPI_SCCB_I2C_PORT,
-            .scl_pin   = EXAMPLE_SPI_SCCB_I2C_SCL_PIN,
-            .sda_pin   = EXAMPLE_SPI_SCCB_I2C_SDA_PIN,
-        },
+            .init_sccb = true,
+            .i2c_config = {
+                .port      = EXAMPLE_SPI_CAM_0_SCCB_I2C_PORT,
+                .scl_pin   = EXAMPLE_SPI_CAM_0_SCCB_I2C_SCL_PIN,
+                .sda_pin   = EXAMPLE_SPI_CAM_0_SCCB_I2C_SDA_PIN,
+            },
 #endif /* !CONFIG_EXAMPLE_SCCB_I2C_INIT_BY_APP */
-        .freq      = EXAMPLE_SPI_SCCB_I2C_FREQ,
-    },
-    .reset_pin = EXAMPLE_SPI_CAM_SENSOR_RESET_PIN,
-    .pwdn_pin  = EXAMPLE_SPI_CAM_SENSOR_PWDN_PIN,
+            .freq      = EXAMPLE_SPI_CAM_0_SCCB_I2C_FREQ,
+        },
+        .reset_pin = EXAMPLE_SPI_CAM_0_SENSOR_RESET_PIN,
+        .pwdn_pin  = EXAMPLE_SPI_CAM_0_SENSOR_PWDN_PIN,
 
-    .spi_port = EXAMPLE_SPI_CAM_SPI_PORT,
-    .spi_cs_pin = EXAMPLE_SPI_CAM_CS_PIN,
-    .spi_sclk_pin = EXAMPLE_SPI_CAM_SCLK_PIN,
-    .spi_data0_io_pin = EXAMPLE_SPI_CAM_DATA0_IO_PIN,
+        .spi_port = EXAMPLE_SPI_CAM_0_SPI_PORT,
+        .spi_cs_pin = EXAMPLE_SPI_CAM_0_CS_PIN,
+        .spi_sclk_pin = EXAMPLE_SPI_CAM_0_SCLK_PIN,
+        .spi_data0_io_pin = EXAMPLE_SPI_CAM_0_DATA0_IO_PIN,
 
-    .xclk_source = EXAMPLE_SPI_CAM_XCLK_RESOURCE,
-    .xclk_pin = EXAMPLE_SPI_CAM_XCLK_PIN,
-    .xclk_freq = EXAMPLE_SPI_CAM_XCLK_FREQ,
+        .xclk_source = EXAMPLE_SPI_CAM_0_XCLK_RESOURCE,
+        .xclk_pin = EXAMPLE_SPI_CAM_0_XCLK_PIN,
+        .xclk_freq = EXAMPLE_SPI_CAM_0_XCLK_FREQ,
 #if CONFIG_EXAMPLE_SPI_CAM_XCLK_USE_LEDC
-    .xclk_ledc_cfg = {
-        .timer = EXAMPLE_SPI_CAM_XCLK_TIMER,
-        .clk_cfg = LEDC_AUTO_CLK,
-        .channel = EXAMPLE_SPI_CAM_XCLK_TIMER_CHANNEL,
-    },
+        .xclk_ledc_cfg = {
+            .timer = EXAMPLE_SPI_CAM_0_XCLK_TIMER,
+            .clk_cfg = LEDC_AUTO_CLK,
+            .channel = EXAMPLE_SPI_CAM_0_XCLK_TIMER_CHANNEL,
+        },
 #endif /* CONFIG_EXAMPLE_SPI_CAM_XCLK_USE_LEDC */
+    },
+
+#if EXAMPLE_ENABLE_SPI_CAM_1_SENSOR
+    {
+        .sccb_config = {
+#if !CONFIG_EXAMPLE_SCCB_I2C_INIT_BY_APP
+            .init_sccb = true,
+            .i2c_config = {
+                .port = EXAMPLE_SPI_CAM_1_SCCB_I2C_PORT,
+                .scl_pin = EXAMPLE_SPI_CAM_1_SCCB_I2C_SCL_PIN,
+                .sda_pin = EXAMPLE_SPI_CAM_1_SCCB_I2C_SDA_PIN,
+            },
+#endif /* !CONFIG_EXAMPLE_SCCB_I2C_INIT_BY_APP */
+            .freq = EXAMPLE_SPI_CAM_1_SCCB_I2C_FREQ,
+        },
+        .reset_pin = EXAMPLE_SPI_CAM_1_SENSOR_RESET_PIN,
+        .pwdn_pin = EXAMPLE_SPI_CAM_1_SENSOR_PWDN_PIN,
+
+        .spi_port = EXAMPLE_SPI_CAM_1_SPI_PORT,
+        .spi_cs_pin = EXAMPLE_SPI_CAM_1_CS_PIN,
+        .spi_sclk_pin = EXAMPLE_SPI_CAM_1_SCLK_PIN,
+        .spi_data0_io_pin = EXAMPLE_SPI_CAM_1_DATA0_IO_PIN,
+
+        .xclk_source = EXAMPLE_SPI_CAM_1_XCLK_RESOURCE,
+        .xclk_pin = EXAMPLE_SPI_CAM_1_XCLK_PIN,
+        .xclk_freq = EXAMPLE_SPI_CAM_1_XCLK_FREQ,
+#if CONFIG_EXAMPLE_SPI_CAM_1_XCLK_USE_LEDC
+        .xclk_ledc_cfg = {
+            .timer = EXAMPLE_SPI_CAM_1_XCLK_TIMER,
+            .clk_cfg = LEDC_AUTO_CLK,
+            .channel = EXAMPLE_SPI_CAM_1_XCLK_TIMER_CHANNEL,
+        },
+#endif /* CONFIG_EXAMPLE_SPI_CAM_1_XCLK_USE_LEDC */
+    },
+#endif /* EXAMPLE_ENABLE_SPI_CAM_1_SENSOR */
 };
 #endif /* EXAMPLE_ENABLE_SPI_CAM_SENSOR */
 
@@ -138,7 +175,7 @@ static const esp_video_init_config_t s_cam_config = {
     .dvp      = &s_dvp_config,
 #endif /* EXAMPLE_ENABLE_DVP_CAM_SENSOR */
 #if EXAMPLE_ENABLE_SPI_CAM_SENSOR
-    .spi      = &s_spi_config,
+    .spi      = s_spi_config,
 #endif /* EXAMPLE_ENABLE_SPI_CAM_SENSOR */
 #if EXAMPLE_ENABLE_USB_UVC_CAM_SENSOR
     .usb_uvc  = &s_usb_uvc_config,
@@ -199,11 +236,15 @@ esp_err_t example_video_init(void)
     dvp_config.sccb_config.init_sccb = false;
     dvp_config.sccb_config.i2c_handle = s_i2cbus_handle;
 #endif /* EXAMPLE_ENABLE_DVP_CAM_SENSOR */
-#if EXAMPLE_ENABLE_SPI_CAM_SENSOR
-    esp_video_init_spi_config_t spi_config = s_spi_config;
-    spi_config.sccb_config.init_sccb = false;
-    spi_config.sccb_config.i2c_handle = s_i2cbus_handle;
-#endif /* EXAMPLE_ENABLE_SPI_CAM_SENSOR */
+#if EXAMPLE_ENABLE_SPI_CAM_0_SENSOR
+    esp_video_init_spi_config_t spi_config[ESP_VIDEO_SPI_DEVICE_NUM];
+
+    memcpy(spi_config, s_spi_config, sizeof(esp_video_init_spi_config_t) * ESP_VIDEO_SPI_DEVICE_NUM);
+    for (int i = 0; i < ESP_VIDEO_SPI_DEVICE_NUM; i++) {
+        spi_config[i].sccb_config.init_sccb = false;
+        spi_config[i].sccb_config.i2c_handle = s_i2cbus_handle;
+    }
+#endif /* EXAMPLE_ENABLE_SPI_CAM_0_SENSOR */
 
     esp_video_init_config_t cam_config = {
 #if EXAMPLE_ENABLE_MIPI_CSI_CAM_SENSOR
@@ -216,7 +257,7 @@ esp_err_t example_video_init(void)
         .dvp      = &dvp_config,
 #endif /* EXAMPLE_ENABLE_DVP_CAM_SENSOR */
 #if EXAMPLE_ENABLE_SPI_CAM_SENSOR
-        .spi      = &spi_config,
+        .spi      = spi_config,
 #endif /* EXAMPLE_ENABLE_SPI_CAM_SENSOR */
 #if EXAMPLE_ENABLE_USB_UVC_CAM_SENSOR
         .usb_uvc  = &s_usb_uvc_config,
@@ -256,13 +297,20 @@ esp_err_t example_video_init(void)
              EXAMPLE_DVP_SCCB_I2C_FREQ);
 #endif /* EXAMPLE_ENABLE_DVP_CAM_SENSOR */
 
-#if EXAMPLE_ENABLE_SPI_CAM_SENSOR
-    ESP_LOGI(TAG, "SPI camera sensor I2C port=%d, scl_pin=%d, sda_pin=%d, freq=%d",
-             EXAMPLE_SPI_SCCB_I2C_PORT,
-             EXAMPLE_SPI_SCCB_I2C_SCL_PIN,
-             EXAMPLE_SPI_SCCB_I2C_SDA_PIN,
-             EXAMPLE_SPI_SCCB_I2C_FREQ);
-#endif /* EXAMPLE_ENABLE_SPI_CAM_SENSOR */
+#if EXAMPLE_ENABLE_SPI_CAM_0_SENSOR
+    ESP_LOGI(TAG, "SPI camera sensor 0 I2C port=%d, scl_pin=%d, sda_pin=%d, freq=%d",
+             EXAMPLE_SPI_CAM_0_SCCB_I2C_PORT,
+             EXAMPLE_SPI_CAM_0_SCCB_I2C_SCL_PIN,
+             EXAMPLE_SPI_CAM_0_SCCB_I2C_SDA_PIN,
+             EXAMPLE_SPI_CAM_0_SCCB_I2C_FREQ);
+#if EXAMPLE_ENABLE_SPI_CAM_1_SENSOR
+    ESP_LOGI(TAG, "SPI camera sensor 1 I2C port=%d, scl_pin=%d, sda_pin=%d, freq=%d",
+             EXAMPLE_SPI_CAM_1_SCCB_I2C_PORT,
+             EXAMPLE_SPI_CAM_1_SCCB_I2C_SCL_PIN,
+             EXAMPLE_SPI_CAM_1_SCCB_I2C_SDA_PIN,
+             EXAMPLE_SPI_CAM_1_SCCB_I2C_FREQ);
+#endif /* EXAMPLE_ENABLE_SPI_CAM_1_SENSOR */
+#endif /* EXAMPLE_ENABLE_SPI_CAM_0_SENSOR */
 
     ESP_GOTO_ON_ERROR(esp_video_init(cam_config_ptr), failed_2, TAG, "failed to initialize video");