+++ esp_video/examples/common_components/example_video_common/include/boards/esp32-s3-eye/example_video_common_board.h
@@ -50,27 +50,27 @@ extern "C" {
 #define EXAMPLE_DVP_D7_PIN                              16
 #endif /* CONFIG_EXAMPLE_ENABLE_DVP_CAM_SENSOR */
 
-#if CONFIG_EXAMPLE_ENABLE_SPI_CAM_SENSOR
+#if CONFIG_EXAMPLE_ENABLE_SPI_CAM_0_SENSOR
 /**
  * @brief SPI camera sensor configuration
  */
-#define EXAMPLE_SPI_SCCB_I2C_SCL_PIN                    5
-#define EXAMPLE_SPI_SCCB_I2C_SDA_PIN                    4
-#define EXAMPLE_SPI_CAM_SENSOR_RESET_PIN                -1
-#define EXAMPLE_SPI_CAM_SENSOR_PWDN_PIN                 -1
+#define EXAMPLE_SPI_CAM_0_SCCB_I2C_SCL_PIN              5
+#define EXAMPLE_SPI_CAM_0_SCCB_I2C_SDA_PIN              4
+#define EXAMPLE_SPI_CAM_0_SENSOR_RESET_PIN              -1
+#define EXAMPLE_SPI_CAM_0_SENSOR_PWDN_PIN               -1
 
 /**
  * @brief SPI camera sensor configuration
  */
-#define EXAMPLE_SPI_CAM_CS_PIN                          6
-#define EXAMPLE_SPI_CAM_SCLK_PIN                        13
-#define EXAMPLE_SPI_CAM_DATA0_IO_PIN                    16
+#define EXAMPLE_SPI_CAM_0_CS_PIN                        6
+#define EXAMPLE_SPI_CAM_0_SCLK_PIN                      13
+#define EXAMPLE_SPI_CAM_0_DATA0_IO_PIN                  16
 
 /**
  * @brief SPI camera sensor clock resource configuration
  */
-#define EXAMPLE_SPI_CAM_XCLK_PIN                        15
-#endif /* CONFIG_EXAMPLE_ENABLE_SPI_CAM_SENSOR */
+#define EXAMPLE_SPI_CAM_0_XCLK_PIN                      15
+#endif /* CONFIG_EXAMPLE_ENABLE_SPI_CAM_0_SENSOR */
 
 #ifdef __cplusplus
 }